Revert "User microposts - issue1 (#2)"
This reverts commit 3b8519824c75a1b9bb5270c7b65f83d6c6423d1f.
