+++ Gemfile
@@ -5,6 +5,8 @@ ruby '3.0.1'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
 gem 'rails', '~> 6.1.4', '>= 6.1.4.1'
+# setup authenticate
+gem 'bcrypt', '~> 3.1.7'
 # Use postgresql as the database for Active Record
 gem 'pg', '~> 1.1'
 # Use Puma as the app server
@@ -19,20 +21,8 @@ gem 'turbolinks', '~> 5'
 gem 'jbuilder', '~> 2.7'
 # Use Redis adapter to run Action Cable in production
 # gem 'redis', '~> 4.0'
-
 # Use Active Model has_secure_password
-gem 'bcrypt', '~> 3.1.7'
-gem 'omniauth'
-gem 'omniauth-rails_csrf_protection'
-gem 'omniauth-google-oauth2'
-gem 'omniauth-facebook'
-
-
-gem 'carrierwave', '~> 2.0'
-
-gem 'mini_magick', '~> 4.5', '>= 4.5.1'
-
-gem 'fog', '~> 1.38'
+# gem 'bcrypt', '~> 3.1.7'
 
 # Use Active Storage variant
 # gem 'image_processing', '~> 1.2'
@@ -46,7 +36,6 @@ gem 'will_paginate-bootstrap', '~> 1.0', '>= 1.0.1'
 gem 'bootsnap', '>= 1.4.4', require: false
 
 group :development, :test do
-  gem 'dotenv-rails'
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
 end
