+++ app/controllers/sessions_controller.rb
@@ -26,29 +26,4 @@ class SessionsController < ApplicationController
     log_out if logged_in?
     redirect_to root_url
   end
-
-  def omniauth #log users in with omniauth
-    user  = User.find_or_create_by(email: auth[:info][:email]) do |u|
-          u.email = auth[:info][:email]
-          u.name = auth[:info][:name]
-          u.uid = auth[:uid]
-          u.provider = auth[:provider]
-          u.password = SecureRandom.hex(10)
-    end
-
-    if user.valid?
-        session[:user_id] = user.id;
-        log_in user
-        redirect_back_or user
-    else
-      flash[:message] = 'Credential error'
-      redirect_to login_path
-    end
-  end
-  
-  private 
-  def auth
-    request.env['omniauth.auth']
-  end
 end
-
