+++ app/controllers/users_controller.rb
@@ -8,7 +8,6 @@ class UsersController < ApplicationController
 
   def show
       @user = User.find(params[:id]);
-      @microposts = @user.microposts.paginate(page: params[:page], :per_page => 10)
   end
 
   def new
@@ -45,12 +44,22 @@ class UsersController < ApplicationController
     flash[:success] = "User deleted"
     redirect_to users_url
   end
+
   private
 
     def user_params
       params.require(:user).permit(:name, :email, :password,
                                   :password_confirmation)
     end 
+
+    # Confirms a logged-in user.
+    def logged_in_user
+      unless logged_in?
+        store_location
+        flash[:danger] = "Please log in."
+        redirect_to login_url
+      end
+    end
 
     # Confirms the correct user.
     def correct_user
