--- app/helpers/microposts_helper.rb
@@ -1,2 +1,0 @@
-module MicropostsHelper
-end
