+++ app/javascript/packs/application.js
@@ -14,3 +14,5 @@ import "./stylesheets/custom.scss"
 Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
+
+
