+++ app/models/user.rb
@@ -1,5 +1,4 @@
 class User < ApplicationRecord
-    has_many :microposts, dependent: :destroy
     attr_accessor :remember_token, :activation_token, :reset_token
     before_save   :downcase_email
     before_create :create_activation_digest
@@ -32,7 +31,7 @@ class User < ApplicationRecord
     end
 
     # Returns true if the given token matches the digest.
-    def authenticated?(attribute, token)
+    def authenticated?(remember_token)
         digest = send("#{attribute}_digest")
         return false if digest.nil?
         BCrypt::Password.new(digest).is_password?(token)
@@ -57,12 +56,6 @@ class User < ApplicationRecord
     # Returns true if a password reset has expired.
     def password_reset_expired?
         reset_sent_at < 2.hours.ago
-    end
-
-    # Defines a proto-feed.
-    # See "Following users" for the full implementation.
-    def feed
-        Micropost.where("user_id = ?", id)
     end
 
     private
