+++ app/views/layouts/_header.html.erb
@@ -28,7 +28,7 @@
               <div class="dropdown-menu mr-5" aria-labelledby="navbarDropdown">
                 <%= link_to "Profile", current_user, class: "dropdown-item" %>
                 <%= link_to "Settings", edit_user_path(current_user) , class: "dropdown-item" %>
-                <%= button_to "Log out", logout_path, method: :delete , class: "dropdown-item"%>
+                <%= link_to "Log out", logout_path, class: "dropdown-item", method: "delete" %>
               </div>
             </li>
           <% else %>
