+++ app/views/sessions/new.html.erb
@@ -11,7 +11,7 @@
       <%= f.label :password %>
       <%= f.password_field :password, class: 'form-control' %>
       <%= link_to "(forgot password)", new_password_reset_path %>
-      <br/>
+
       <%= f.label :remember_me, class: "checkbox inline" do %>
         <%= f.check_box :remember_me %>
         <span>Remember me on this computer</span>
@@ -19,8 +19,7 @@
 
       <%= f.submit "Log in", class: "btn btn-primary" %>
     <% end %>
-    <%= button_to "Login with google", "/auth/google_oauth2", method: :post,  class: "btn btn-primary" %>
-    <%= button_to "Login with facebook", "/auth/facebook", method: :post,  class: "btn btn-primary" %>
+
     <p>New user? <%= link_to "Sign up now!", signup_path %></p>
   </div>
 </div>
