+++ app/views/user_mailer/account_activation.html.erb
@@ -1,10 +1,5 @@
-<h1>Sample App</h1>
-
-<p>Hi <%= @user.name %>,</p>
+<h1>User#account_activation</h1>
 
 <p>
-Welcome to the Sample App! Click on the link below to activate your account:
+  <%= @greeting %>, find me in app/views/user_mailer/account_activation.html.erb
 </p>
-
-<%= link_to "Activate", edit_account_activation_url(@user.activation_token,
-                                                    email: @user.email) %>
