+++ app/views/users/new.html.erb
@@ -4,7 +4,7 @@
 <div class="row justify-content-center">
   <div class="col-md-6 col-md-offset-3">
     <%= form_for(@user) do |f| %>
-      <%= render 'shared/error_messages' , object: f.object%>
+      <%= render 'shared/error_messages' %>
       <%= f.label :name %>
       <%= f.text_field :name, class: 'form-control'%>
