--- config/initializers/omniauth.rb
@@ -1,4 +1,0 @@
-Rails.application.config.middleware.use OmniAuth::Builder do
-    provider :google_oauth2, ENV['GOOGLE_CLIENT_ID'], ENV['GOOGLE_CLIENT_SECRET']
-    provider :facebook, ENV['FACEBOOK_CLIENT_ID'], ENV['FACEBOOK_CLIENT_SECRET']
-end
