+++ config/routes.rb
@@ -13,8 +13,5 @@ Rails.application.routes.draw do
   resources :users
   resources :account_activations, only: [:edit]
   resources :password_resets,     only: [:new, :create, :edit, :update]
-  resources :microposts,          only: [:create, :destroy]
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-  get '/auth/:provider/callback', to: 'sessions#omniauth'
-  
 end
