--- db/migrate/20211123074909_add_associate_to_users.rb
@@ -1,7 +1,0 @@
-class AddAssociateToUsers < ActiveRecord::Migration[6.1]
-  def change
-    add_column :users, :provider, :string
-    add_column :users, :uid, :string
-    add_column :users, :image, :string
-  end
-end
