--- db/migrate/20211127034012_create_microposts.rb
@@ -1,11 +1,0 @@
-class CreateMicroposts < ActiveRecord::Migration[6.1]
-  def change
-    create_table :microposts do |t|
-      t.text :content
-      t.references :user, null: false, foreign_key: true
-
-      t.timestamps
-    end
-    add_index :microposts, [:user_id, :created_at]
-  end
-end
