--- db/migrate/20211127134358_add_picture_to_microposts.rb
@@ -1,5 +1,0 @@
-class AddPictureToMicroposts < ActiveRecord::Migration[6.1]
-  def change
-    add_column :microposts, :picture, :string
-  end
-end
