--- test/fixtures/microposts.yml
@@ -1,9 +1,0 @@
-# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
-
-one:
-  content: MyText
-  user: one
-
-two:
-  content: MyText
-  user: two
